use DefaultClusterStore
