+++ apps/portal-web/src/layouts/BaseLayout.tsx
@@ -15,8 +15,8 @@ import { BaseLayout as LibBaseLayout } from "@scow/lib-web/build/layouts/base/Ba
 import { JumpToAnotherLink } from "@scow/lib-web/build/layouts/base/header/components";
 import { PropsWithChildren, useMemo } from "react";
 import { useStore } from "simstate";
-import { useDefaultCluster } from "src/layouts/DefaultCluster";
 import { userRoutes } from "src/layouts/routes";
+import { DefaultClusterStore } from "src/stores/DefaultClusterStore";
 import { LoginNodeStore } from "src/stores/LoginNodeStore";
 import { UserStore } from "src/stores/UserStore";
 import { publicConfig } from "src/utils/config";
@@ -30,7 +30,7 @@ export const BaseLayout = ({ footerText, versionTag, children }: PropsWithChildr
 
   const userStore = useStore(UserStore);
   const loginNodes = useStore(LoginNodeStore);
-  const { defaultCluster, setDefaultCluster, removeDefaultCluster } = useDefaultCluster();
+  const { defaultCluster, setDefaultCluster, removeDefaultCluster } = useStore(DefaultClusterStore);
 
   const routes = useMemo(() => userRoutes(
     userStore.user, defaultCluster, loginNodes, setDefaultCluster,
