+++ apps/portal-web/src/pageComponents/job/AllJobsTable.tsx
@@ -20,9 +20,11 @@ import Router from "next/router";
 import { join } from "path";
 import React, { useCallback, useMemo, useState } from "react";
 import { useAsync } from "react-async";
+import { useStore } from "simstate";
 import { api } from "src/apis";
+import { SingleClusterSelector } from "src/components/ClusterSelector";
 import { FilterFormContainer } from "src/components/FilterFormContainer";
-import { SingleClusterSelector, useDefaultCluster } from "src/layouts/DefaultCluster";
+import { DefaultClusterStore } from "src/stores/DefaultClusterStore";
 import { Cluster } from "src/utils/config";
 
 interface FilterForm {
@@ -39,7 +41,7 @@ export const AllJobQueryTable: React.FC<Props> = ({
   userId,
 }) => {
 
-  const { defaultCluster } = useDefaultCluster();
+  const { defaultCluster } = useStore(DefaultClusterStore);
 
   const [query, setQuery] = useState<FilterForm>(() => {
     const now = dayjs();
