+++ apps/portal-web/src/pageComponents/job/JobTemplateTable.tsx
@@ -15,9 +15,11 @@ import { Button, Form, Space, Table } from "antd";
 import Link from "next/link";
 import React, { useCallback, useState } from "react";
 import { useAsync } from "react-async";
+import { useStore } from "simstate";
 import { api } from "src/apis";
+import { SingleClusterSelector } from "src/components/ClusterSelector";
 import { FilterFormContainer } from "src/components/FilterFormContainer";
-import { SingleClusterSelector, useDefaultCluster } from "src/layouts/DefaultCluster";
+import { DefaultClusterStore } from "src/stores/DefaultClusterStore";
 import type { Cluster } from "src/utils/config";
 
 interface Props {}
@@ -28,7 +30,7 @@ interface FilterForm {
 
 export const JobTemplateTable: React.FC<Props> = () => {
 
-  const { defaultCluster } = useDefaultCluster();
+  const { defaultCluster } = useStore(DefaultClusterStore);
 
   const [query, setQuery] = useState<FilterForm>(() => {
     return {
