+++ apps/portal-web/src/pages/_app.tsx
@@ -32,8 +32,8 @@ import { USE_MOCK } from "src/apis/useMock";
 import { getTokenFromCookie } from "src/auth/cookie";
 import { AntdConfigProvider } from "src/layouts/AntdConfigProvider";
 import { BaseLayout } from "src/layouts/BaseLayout";
-import { DefaultClusterProvider } from "src/layouts/DefaultCluster";
 import { FloatButtons } from "src/layouts/FloatButtons";
+import { DefaultClusterStore } from "src/stores/DefaultClusterStore";
 import { LoginNodeStore } from "src/stores/LoginNodeStore";
 import {
   User, UserStore,
@@ -103,6 +103,8 @@ function MyApp({ Component, pageProps, extra }: Props) {
 
   const loginNodeStore = useConstant(() => createStore(LoginNodeStore, extra.loginNodes));
 
+  const defaultClusterStore = useConstant(() => createStore(DefaultClusterStore));
+
   // Use the layout defined at the page level, if available
   return (
     <>
@@ -126,18 +128,16 @@ function MyApp({ Component, pageProps, extra }: Props) {
           }}
         />
       </Head>
-      <StoreProvider stores={[userStore, loginNodeStore]}>
+      <StoreProvider stores={[userStore, defaultClusterStore, loginNodeStore]}>
         <DarkModeProvider initial={extra.darkModeCookieValue}>
           <AntdConfigProvider color={primaryColor}>
             <FloatButtons />
             <GlobalStyle />
             <FailEventHandler />
             <TopProgressBar />
-            <DefaultClusterProvider>
-              <BaseLayout footerText={footerText} versionTag={publicConfig.VERSION_TAG}>
-                <Component {...pageProps} />
-              </BaseLayout>
-            </DefaultClusterProvider>
+            <BaseLayout footerText={footerText} versionTag={publicConfig.VERSION_TAG}>
+              <Component {...pageProps} />
+            </BaseLayout>
           </AntdConfigProvider>
         </DarkModeProvider>
       </StoreProvider>
